Moved Data objects to DTO package
This change has a mild effect on class organization and discovery by Spring, not strictly necessary but definitely helpful
